Удалил общий контроллер для двух видов визиток, разделил на два разных окна Мои визитки и Контакты, ускорил работу окна Настроек, добавил кнопку Отменить в окне Создания визиток
